+++ fsic.py
@@ -5,7 +5,7 @@ fsic
 Tools for macroeconomic modelling in Python.
 """
 
-__version__ = '0.5.0.dev'
+__version__ = '0.5.1.dev'
 
 
 import copy
@@ -753,6 +753,23 @@ class BaseModel:
 
         # Optionally copy initial values from another period
         if offset:
+
+            # Error if `offset` points prior to the current model span
+            if t + offset < 0:
+                raise IndexError(
+                    '`offset` argument ({}) for position `t` ({}) '
+                    'implies a period before the span of the current model instance: '
+                    '{} + {} = position {} < 0'.format(
+                        offset, t, offset, t, offset + t))
+
+            # Error if `offset` points beyond the current model span
+            if t + offset >= len(self.span):
+                raise IndexError(
+                    '`offset` argument ({}) for position `t` ({}) '
+                    'implies a period beyond the span of the current model instance: '
+                    '{} + {} = position {} >= {} periods in span'.format(
+                        offset, t, offset, t, offset + t, len(self.span)))
+
             for name in self.ENDOGENOUS:
                 self.__dict__['_' + name][t] = self.__dict__['_' + name][t + offset]